backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
pydantic==1.10.7
python-multipart==0.0.6
redis==4.5.5
websockets==11.0.3
pandas==2.0.1
numpy==1.24.3
scikit-learn==1.2.2
scikit-surprise==1.1.3
lightfm==1.17
xgboost==1.7.5
stable-baselines3==1.8.0
gymnasium==0.28.1
tensorflow==2.12.0
recsim==0.2.4
python-dotenv==1.0.0
pytest==7.3.1
httpx==0.24.0
alembic==1.10.4
bcrypt==4.0.1
email-validator==2.0.0